Reorder the elements in import
As happens in:

CSS -> Shorthand notation
CSS -> Nesting in Less and Sass

First appears the bad way and next the good way to do it. So I think made more sense to have a uniform in all elements, first the way to avoid and after the way how should be do it.